Add files via upload
The main goal of the Blood Bank Management System project is to monitor Blood Bank data, Blood stock, Donor List, and Patient List. It manages all the Blood Bank, Donor, Blood stock data, and Patient Details. The project is entirely administrative and therefore access is guaranteed only to the administrator.